# api/gen/api.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package gen

import (
	"time"
)

// Leaderboard defines model for Leaderboard.
type Leaderboard struct {
	CorrectResults   int    `json:"correctResults"`
	CorrectScores    int    `json:"correctScores"`
	JokersPlayed     int    `json:"jokersPlayed"`
	Participant      string `json:"participant"`
	Played           int    `json:"played"`
	PointsDifference int    `json:"pointsDifference"`
	Position         int    `json:"position"`
	PreviousPosition int    `json:"previousPosition"`
	TotalPoints      int    `json:"totalPoints"`
}

// Match defines model for Match.
type Match struct {
	AwayScore *int      `json:"awayScore,omitempty"`
	AwayTeam  string    `json:"awayTeam"`
	Date      time.Time `json:"date"`
	HasResult bool      `json:"hasResult"`
	HomeScore *int      `json:"homeScore,omitempty"`
	HomeTeam  string    `json:"homeTeam"`
	Id        int       `json:"id"`
	Round     string    `json:"round"`
}

// MatchNavigation defines model for MatchNavigation.
type MatchNavigation struct {
	AwayTeam string `json:"awayTeam"`
	HomeTeam string `json:"homeTeam"`
	Id       int    `json:"id"`
}

// MatchPredictions defines model for MatchPredictions.
type MatchPredictions struct {
	Id          int              `json:"id"`
	Match       Match            `json:"match"`
	NextNav     *MatchNavigation `json:"nextNav,omitempty"`
	Predictions []Prediction     `json:"predictions"`
	PreviousNav *MatchNavigation `json:"previousNav,omitempty"`
}

// Participant defines model for Participant.
type Participant struct {
	Name                  string                `json:"name"`
	Predictions           []Prediction          `json:"predictions"`
	TotalPoints           int                   `json:"totalPoints"`
	TournamentPredictions TournamentPredictions `json:"tournamentPredictions"`
}

// ParticipantPrediction defines model for ParticipantPrediction.
type ParticipantPrediction struct {
	AwayScore   int    `json:"awayScore"`
	HomeScore   int    `json:"homeScore"`
	MatchId     int    `json:"matchId"`
	Participant string `json:"participant"`
	PlayedJoker *bool  `json:"playedJoker,omitempty"`
}

// Prediction defines model for Prediction.
type Prediction struct {
	AwayScore   int    `json:"awayScore"`
	HasResult   bool   `json:"hasResult"`
	HomeScore   int    `json:"homeScore"`
	Id          int    `json:"id"`
	Participant string `json:"participant"`
	Points      int    `json:"points"`
	UsedJoker   *bool  `json:"usedJoker,omitempty"`
}

// Team defines model for Team.
type Team struct {
	Code        string `json:"code"`
	DisplayName string `json:"displayName"`
}

// TournamentPredictions defines model for TournamentPredictions.
type TournamentPredictions struct {
	FourthPlace       string `json:"fourthPlace"`
	RunnerUp          string `json:"runnerUp"`
	ScorerNationality string `json:"scorerNationality"`
	ThirdPlace        string `json:"thirdPlace"`
	TopScorer         string `json:"topScorer"`
	Winner            string `json:"winner"`
}

// GetPredictionsParams defines parameters for GetPredictions.
type GetPredictionsParams struct {
	MatchId     *int    `form:"matchId,omitempty" json:"matchId,omitempty"`
	Participant *string `form:"participant,omitempty" json:"participant,omitempty"`
}

// CreatePredictionJSONRequestBody defines body for CreatePrediction for application/json ContentType.
type CreatePredictionJSONRequestBody = ParticipantPrediction