# api/openapi.yaml
openapi: 3.0.3
info:
  title: World Cup Predictions API
  version: 1.0.0
components:
  schemas:
...
    Prediction:
      type: object
      required:
        - id
        - participant
        - points
        - homeScore
        - awayScore
        - hasResult
      properties:
        id:
          type: integer

        participant:
          type: string

        points:
          type: integer

        homeScore:
          type: integer

        awayScore:
          type: integer

        usedJoker:
          type: boolean

        hasResult:
          type: boolean
...
    ParticipantPrediction:
      type: object
      required:
        - participant
        - matchId
        - homeScore
        - awayScore

      properties:
        participant:
          type: string

        matchId:
          type: integer

        homeScore:
          type: integer

        awayScore:
          type: integer

        playedJoker:
          type: boolean
paths:
...
  /api/predictions:
    get:
      operationId: getPredictions
      summary: Get predictions (optionally filtered for match or participant)
      parameters:
        - name: matchId
          in: query
          required: false
          schema:
            type: integer
        - name: participant
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prediction"
          
    post:
      operationId: createPrediction
      summary: Enter a prediction for a participant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParticipantPrediction"
      responses:
        "201":
          description: Prediction Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipantPrediction"